requirements.txt
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.21.4
opencv-python==4.5.3.56
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
toml==0.10.2
torchaudio==0.9.1
tqdm==4.62.3
pyyaml

